requirements.txt
certifi==2024.7.4
charset-normalizer==3.3.2
clip @ git+https://github.com/openai/CLIP.git@dcba3cb2e2827b402d2701e7e1c7d9fed8a20ef1
filelock==3.15.4
fsspec==2024.6.1
ftfy==6.2.0
huggingface-hub==0.24.2
idna==3.7
Jinja2==3.1.4
joblib==1.4.2
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.3
numpy==1.26.0
open_clip_torch==2.26.1
opencv-python==4.10.0.84
packaging==24.1
pillow==10.4.0
PyYAML==6.0.1
regex==2024.7.24
requests==2.32.3
safetensors==0.4.3
scikit-learn==1.5.1
scipy==1.14.0
sentence-transformers==3.0.1
sympy==1.13.1
threadpoolctl==3.5.0
timm==1.0.8
tokenizers==0.19.1
torch==2.2.2
torchvision==0.17.2
tqdm==4.66.4
transformers==4.43.3
typing_extensions==4.12.2
urllib3==2.2.2
wcwidth==0.2.13
